feat: move car file import into the local coreapi implementation
